# PctBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Heures UVCI API",
        "description": "Documentation de l'API de gestion des heures",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/v1/parametres": {
            "get": {
                "tags": [
                    "Paramètres de calcul"
                ],
                "summary": "Liste de tous les paramètres de calcul",
                "operationId": "b6297fcef4ca042943cf2be23cef0118",
                "responses": {
                    "200": {
                        "description": "Liste des paramètres"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Paramètres de calcul"
                ],
                "summary": "Créer un paramètre de calcul",
                "operationId": "a4b720f2a9c920fea9516b46d4d1f9d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type_operation",
                                    "niveau",
                                    "libelle_niveau",
                                    "heures_par_seance",
                                    "id_annee"
                                ],
                                "properties": {
                                    "type_operation": {
                                        "type": "string",
                                        "enum": [
                                            "conception",
                                            "mise_a_jour"
                                        ]
                                    },
                                    "niveau": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "libelle_niveau": {
                                        "type": "string",
                                        "example": "Simple"
                                    },
                                    "heures_par_seance": {
                                        "type": "number",
                                        "example": 1.5
                                    },
                                    "id_annee": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Taux pour conception simple"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paramètre créé"
                    },
                    "422": {
                        "description": "Ce paramètre existe déjà"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/parametres/{id}": {
            "get": {
                "tags": [
                    "Paramètres de calcul"
                ],
                "summary": "Détail d'un paramètre",
                "operationId": "9deeb28c7ccaf273bab5d4982d67b3e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail paramètre"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Paramètres de calcul"
                ],
                "summary": "Modifier le coefficient d'un paramètre",
                "operationId": "5cd5e245baa67c70e60843b46506f2ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "coefficient_vhn": {
                                        "type": "number",
                                        "example": 5
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paramètre modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Paramètres de calcul"
                ],
                "summary": "Supprimer un paramètre de calcul",
                "operationId": "52bb59485ca925d46b66e4e75bc61db5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paramètre supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/rapports/enseignant/{id}": {
            "get": {
                "tags": [
                    "Rapports"
                ],
                "summary": "Fiche individuelle enseignant (PDF)",
                "operationId": "18043aeb302acf8d3561eead1f115a4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_annee",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF généré",
                        "content": {
                            "application/pdf": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}